pyproject.toml
[tool.poetry]
name = "python-for-sql"
version = "0.0.1"
description = "python package for interacting with sql databases"
authors = ["Louis"]
readme = "README.md"
packages = [{include = "python_for_sql"}]

[tool.poetry.dependencies]
python = "3.10"
pandas = "^2.0.2"
psycopg2 = "^2.9.6"
SQLAlchemy = "^2.0.17"
python-dotenv = "^1.0.0"
ipykernel = "^6.23.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
